scripts/diarization-poc/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "diarization-poc"
version = "0.1.0"
description = "Standalone diarization and speaker identity learning pipeline"
requires-python = ">=3.10"
dependencies = [
  "numpy==2.1.3",
  "python-dotenv==1.1.1",
  "requests==2.32.3",
  "sounddevice==0.5.1",
  "webrtcvad==2.0.10",
  "resemblyzer==0.1.4",
]

[tool.uv]
package = false
